requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
itsdangerous
webauthn
passlib
authlib
httpx
python-dotenv
